requirements.txt
alembic==1.10.2
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
